# FDX-Core-API.openapi.yaml
openapi: 3.1.0
info:
  version: 6.0.0
  title: FDX Core API
  description: Financial Data Exchange V6.0.0 Core API
  contact:
    name: Financial Data Exchange
    url: 'https://financialdataexchange.org/'
    email: fdxsupport@financialdataexchange.org
  license:
    url: 'https://financialdataexchange.org/common/Uploaded%20files/Policies/FDX%20API%20License%20Agreement-(11-13-2019).pdf'
    name: FDX API License Agreement (11-13-2019)
servers:
  - url: 'https://baseurl/fdx/v6/core'
    description: Financial Data Exchange V6.0.0 Core API
paths:
...
  '/accounts/{accountId}/transactions':
    get:
      tags:
        - Account Transactions
      summary: Search for account transactions
      description: 'Search for account transactions. Example: /accounts/{accountId}/transactions?startTime=value1&endTime=value2'
      operationId: searchForAccountTransactions
      parameters:
        - name: Authorization
          in: header
          description: Holds Bearer information
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/StartTimeQuery'
        - $ref: '#/components/parameters/EndTimeQuery'
        - $ref: '#/components/parameters/AccountIdPath'
        - $ref: '#/components/parameters/OffsetQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/FapiInteractionIdHeader'
        - $ref: '#/components/parameters/FdxApiActorTypeHeader'
      responses:
        '200':
          description: 'Paginated collection of transactions, which can be one of DepositTransaction, LoanTransaction, LineOfCreditTransaction, InvestmentTransaction, InsuranceTransaction, or CommercialTransaction'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transactions'
              examples:
                Paginated Transactions:
                  value:
                    page:
                      nextOffset: '2'
                      total: 3
                    links:
                      next:
                        href: /accounts/33333/transactions?offSet=2&limit=10
                    transactions:
                      - accountCategory: DEPOSIT_ACCOUNT
                        accountId: '10001'
                        transactionId: '20002'
                        postedTimestamp: '2017-11-05T13:15:30.751Z'
                        description: Direct deposit from XYZ Corp
                        debitCreditMemo: CREDIT
                        amount: 1200.42
                        payee: XYZ Corp
                      - accountCategory: DEPOSIT_ACCOUNT
                        accountId: '10001'
                        transactionId: '20002'
                        postedTimestamp: '2017-11-05T13:15:31.751Z'
                        description: Withdrawal from ATM
                        debitCreditMemo: DEBIT
                        amount: 1200.42
                        payee: Account Owner
                Transaction with Rewards:
                  summary: Example of single transaction with associated reward
                  value:
                    transactions:
                      - accountCategory: LOC_ACCOUNT
                        accountId: '10001'
                        transactionId: '20003'
                        postedTimestamp: '2017-11-05T13:15:30.751Z'
                        description: Hotel Stay
                        debitCreditMemo: DEBIT
                        amount: 236
                        reward:
                          accrued: 472
                          adjusted: 0
                          categoryId: '293'
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/x-fapi-interaction-id'
        '400':
          description: Input sent by client does not satisfy API specification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/x-fapi-interaction-id'
        '403':
          description: 'Forbidden, server understands the request but refuses to authorize it'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Inactive TPP:
                  value:
                    code: '602'
                    message: Customer not authorized
                    debugMessage: Authenticated customer does not have the authorization to perform this action.
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/x-fapi-interaction-id'
        '404':
          description: Account with id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Account not found:
                  value:
                    code: '701'
                    message: Account not found
                    debugMessage: Provider custom developer-level error details for troubleshooting
                Data not found for request parameters:
                  value:
                    code: '1107'
                    message: Data not found for request parameters
                    debugMessage: Provider custom developer-level error details for troubleshooting
          headers:
            x-fapi-interaction-id:
              $ref: '#/components/headers/x-fapi-interaction-id'
        '500':
          $ref: '#/components/responses/500'
        '501':
          $ref: '#/components/responses/501'
        '503':
          $ref: '#/components/responses/503'
      servers:
        - url: 'https://baseurl/fdx/v6/core'
          description: Financial Data Exchange V6.0.0 Core API
...
components:
  parameters:
    EndTimeQuery:
      name: endTime
      in: query
      description: End time for use in retrieval of elements (ISO 8601)
      schema:
        $ref: '#/components/schemas/DateString'
    StartTimeQuery:
      name: startTime
      in: query
      description: Start time for use in retrieval of elements (ISO 8601)
      schema:
        $ref: '#/components/schemas/DateString'
    AccountIdPath:
      name: accountId
      in: path
      description: Account Identifier
      required: true
      schema:
        type: string
...
    OffsetQuery:
      name: offset
      in: query
      description: Opaque cursor used by the provider to send the next set of records
      schema:
        type: string
    LimitQuery:
      name: limit
      in: query
      description: Number of elements that the consumer wishes to receive. Providers should implement reasonable default and maximum values
      schema:
        type: integer
    FapiInteractionIdHeader:
      name: x-fapi-interaction-id
      in: header
      description: 'Unique identifier for this interaction as an [IETF RFC 4122](https://datatracker.ietf.org/doc/html/rfc4122) UUID. Used across all FDX API requests and responses'
      required: true
      schema:
        type: string
        examples:
          - c770aef3-6784-41f7-8e0e-ff5f97bddb3a
      examples:
        default:
          value: c770aef3-6784-41f7-8e0e-ff5f97bddb3a
    FdxApiActorTypeHeader:
      name: FDX-API-Actor-Type
      in: header
      description: Identifies whether the customer is present (USER) or it is a BATCH operation
      schema:
        $ref: '#/components/schemas/ActorType'
        example: BATCH
      examples:
        default:
          value: BATCH
  headers:
    x-fapi-interaction-id:
      description: Unique identifier for this interaction
      schema:
        type: string
      example: rrt-970585410023330903-b-gce-821-23397-1
  securitySchemes:
    OAuthFapi1Advanced:
      type: oauth2
      description: |-
        This authorization protocol uses FAPI Security Profile 1.0 Part Two — Advanced. To be used with the FDX API customer data domains. In addition, these operations employ the FDX User Consent framework for permissioned data sharing.
        Client must comply with [5.2.3 Confidential client](https://openid.net/specs/openid-financial-api-part-2-1_0.html#confidential-client). Authorization server must comply with the [5.2.2 Authorization server](https://openid.net/specs/openid-financial-api-part-2-1_0.html#authorization-server).
        Client, authorization server and resource server requirements described in full by:
          * [FAPI Security Profile 1.0 Part Two — Advanced](https://openid.net/specs/openid-financial-api-part-2-1_0.html)
          * FDX API Security Model specification published with the FDX API

        Consent User Experience and Issuance described by:
          * FDX User Experience Guidelines published with the FDX API
          * Consent API operations and data structures published within the FDX API
      flows:
        authorizationCode:
          authorizationUrl: 'https://auth.fi.com/fdx/v6/par'
          scopes:
            'fdx:accountbasic:read': 'Scope authorizing retrieval of account display name, masked account number, type, description'
            'fdx:accountdetailed:read': 'Scope authorizing retrieval of account information. Includes data from "Basic" + account balances, credit limits, due dates, interest rates, rewards balances, recurrences'
            'fdx:accountpayments:read': Scope authorizing retrieval of full account number and payment network routing number
            'fdx:customercontact:read': 'Scope authorizing retrieval of name, email, address, phone on file with this institution. Applies to both user and other account holders'
            'fdx:customerpersonal:read': 'Scope authorizing retrieval of name, email, address, phone, DoB, tax ID, SSN on file with this institution'
            'fdx:investments:read': Scope authorizing retrieval of details about individual underlying investment positions
            'fdx:transactions:read': 'Scope authorizing retrieval of historical and current transactions, transaction types, amounts, dates and descriptions'
            'fdx:statements:read': 'Scope authorizing retrieval of periodic PDF statement showing personal information, account and transaction details. May contain PII such as name, address'
            offline_access: Scope authorizing the client to refresh `access_tokens` without user present
            openid: Scope authorizing access to the user's identity details
          tokenUrl: 'https://auth.fi.com/fdx/v6/token'
  responses:
...
    '500':
      description: Catch all exception where request was not processed due to an internal outage/issue. Consider other more specific errors before using this error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            Internal server error:
              value:
                code: '500'
                message: Internal server error
                debugMessage: Provider custom developer-level error details for troubleshooting
      headers:
        x-fapi-interaction-id:
          $ref: '#/components/headers/x-fapi-interaction-id'
    '501':
      description: Error when FdxVersion in Header is not one of those implemented at backend
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            FdxVersion not supported or not implemented:
              value:
                code: '1106'
                message: FdxVersion not supported or not implemented
                debugMessage: Provider custom developer-level error details for troubleshooting
      headers:
        x-fapi-interaction-id:
          $ref: '#/components/headers/x-fapi-interaction-id'
    '503':
      description: System is down for maintenance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            Scheduled Maintenance:
              value:
                code: '503'
                message: Scheduled Maintenance
                debugMessage: Provider custom developer-level error details for troubleshooting
      headers:
        x-fapi-interaction-id:
          $ref: '#/components/headers/x-fapi-interaction-id'
  schemas:
...
    PaginatedArray:
      type: object
      title: Paginated Array
      description: Base class for results that may be paginated
      properties:
        page:
          $ref: '#/components/schemas/PageMetadata'
          description: Offset IDs for navigating result sets
        links:
          $ref: '#/components/schemas/PageMetadataLinks'
          description: Resource URLs for navigating result sets
    PageMetadata:
      type: object
      title: Page Metadata
      description: Offset IDs for paginated result sets
      properties:
        nextOffset:
          type: string
          description: Opaque identifier. Does not need to be numeric or have any specific pattern. Implementation specific
          examples:
            - '2'
        prevOffset:
          type: string
          description: Opaque identifier. Does not need to be numeric or have any specific pattern. Implementation specific
          examples:
            - '1'
        totalElements:
          type: integer
          description: Total number of elements
          examples:
            - 3
    PageMetadataLinks:
      type: object
      title: Page Metadata Links
      description: Resource URLs for retrieving next or previous datasets
      properties:
        next:
          $ref: '#/components/schemas/HateoasLink'
          description: Resource URL for retrieving next dataset
        prev:
          $ref: '#/components/schemas/HateoasLink'
          description: Resource URL for retrieving previous dataset
...
    Error:
      type: object
      title: Error
      description: An error entity which can be used at the API level for error responses or at the account level to indicate a problem specific to a particular account
      properties:
        code:
          type: string
          description: Long term persistent identifier which can be used to trace error condition back to log information
        message:
          type: string
          description: End user displayable information which might help the customer diagnose an error
        debugMessage:
          type: string
          description: Message used to debug the root cause of the error. Contents should not be used in consumer's business logic. Can change at any time and should only be used for consumer to communicate with the data provider about an issue. Provider can include an error GUID in message for their use.
...
    DateString:
      type: string
      title: Date String
      description: 'ISO 8601 full-date in format ''YYYY-MM-DD'' according to [IETF RFC3339](https://xml2rfc.tools.ietf.org/public/rfc/html/rfc3339.html#anchor14)'
      format: date
      maxLength: 10
      examples:
        - '2021-07-15'
...
    CommercialCode:
      type: object
      title: Commercial Code entity
      description: |-
        The code for a specific treasury management defined field.
        The [X9 BTRS/BTR3 standard](https://x9.org/standards/btrs/) (formerly named BAI2) for corporate and treasury data elements is proposed to provide the basis of (or starting point for) representing summary, account, and transaction codes for the Corp/Treasury Entity, subject to any necessary modifications useful to the members of the FDX consortium and users of the API specification
      properties:
        type:
          $ref: '#/components/schemas/TreasuryManagementType'
        code:
          type: string
          description: The code of the Treasury Management defined field
...
    Inv401kSourceType:
      type: string
      enum:
        - AFTERTAX
        - MATCH
        - OTHERNONVEST
        - OTHERVEST
        - PRETAX
        - PROFITSHARING
        - ROLLOVER
      title: Investment 401k Source Type
      description: The source of a 401k investment
...
    UnitType:
      type: string
      enum:
        - CURRENCY
        - SHARES
      title: Unit Type
      description: The units of an investment transaction
...
    SubAccountType:
      type: string
      enum:
        - CASH
        - MARGIN
        - OTHER
        - SHORT
      title: Sub Account Type
      description: The subtype of an account
...
    Transactions:
      type: object
      title: Transactions entity
      description: Optionally paginated array of transactions
      allOf:
        - $ref: '#/components/schemas/PaginatedArray'
        - type: object
          properties:
            transactions:
              type: array
              description: 'An array of transactions with entity type dependent on the account type (commercial, deposit, insurance, investment, line of credit or loan)'
              items:
                allOf:
                  - $ref: '#/components/schemas/CommercialTransaction'
                  - $ref: '#/components/schemas/DepositTransaction'
                  - $ref: '#/components/schemas/InsuranceTransaction'
                  - $ref: '#/components/schemas/InvestmentTransaction'
                  - $ref: '#/components/schemas/LineOfCreditTransaction'
                  - $ref: '#/components/schemas/LoanTransaction'
                discriminator:
                  propertyName: accountCategory
                  mapping:
                    COMMERCIAL_ACCOUNT: '#/components/schemas/CommercialTransaction'
                    DEPOSIT_ACCOUNT: '#/components/schemas/DepositTransaction'
                    INSURANCE_ACCOUNT: '#/components/schemas/InsuranceTransaction'
                    INVESTMENT_ACCOUNT: '#/components/schemas/InvestmentTransaction'
                    LOAN_ACCOUNT: '#/components/schemas/LoanTransaction'
                    LOC_ACCOUNT: '#/components/schemas/LineOfCreditTransaction'
    CommercialTransaction:
      type: object
      title: Commercial Transaction entity
      description: A transaction on a commercial account type
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            immediateAvailableBalance:
              type: number
              description: Immediate available balance
            nextDayAvailableBalance:
              type: number
              description: Next day available balance
            twoDaysPlusAvailableBalance:
              type: number
              description: Two days plus available balance
            referenceBankId:
              type: string
              description: Reference bank id
            referenceBranchId:
              type: string
              description: Reference branch id
            referenceCustomerId:
              type: string
              description: Reference customer id
            commercialCode:
              $ref: '#/components/schemas/CommercialCode'
              description: The code for a specific treasury management defined field
            memo:
              type: string
              description: Memo field for the commercial transaction
    Transaction:
      type: object
      title: Transaction
      description: Base entity for financial transactions
      properties:
        accountCategory:
          $ref: '#/components/schemas/AccountCategory'
          description: The entity type of the account of this Transaction
        accountId:
          $ref: '#/components/schemas/Identifier'
          description: Corresponds to AccountId in Account
        transactionId:
          $ref: '#/components/schemas/Identifier'
          description: Long term persistent identity of the transaction (unique to account)
        referenceTransactionId:
          $ref: '#/components/schemas/Identifier'
          description: 'For reverse postings, the identity of the transaction being reversed. For the correction transaction, the identity of the reversing post. For credit card posting transactions, the identity of the authorization transaction'
        postedTimestamp:
          $ref: '#/components/schemas/Timestamp'
          description: The date and time that the transaction was posted to the account. If not provided then TransactionTimestamp can be used as PostedTimeStamp
        transactionTimestamp:
          $ref: '#/components/schemas/Timestamp'
          description: The date and time that the transaction was added to the server backend systems
        description:
          type: string
          description: The description of the transaction
        memo:
          $ref: '#/components/schemas/String255'
          description: Secondary transaction description
        debitCreditMemo:
          $ref: '#/components/schemas/DebitCreditMemo'
          description: 'DEBIT, CREDIT, MEMO'
        category:
          type: string
          description: 'Transaction category, preferably MCC or SIC.'
        subCategory:
          type: string
          description: Transaction category detail
        reference:
          type: string
          description: A tracking reference identifier
        status:
          $ref: '#/components/schemas/TransactionStatus'
          description: 'PENDING, MEMO, POSTED, AUTHORIZATION'
        amount:
          type: number
          description: The amount of money in the account currency
        foreignAmount:
          type: number
          description: The amount of money in the foreign currency
        foreignCurrency:
          $ref: '#/components/schemas/Iso4217CurrencyCode'
          description: The ISO 4217 code of the foreign currency
        imageIds:
          type: array
          description: Array of Image Identifiers (unique to this transaction) used to retrieve Images of check or transaction receipt
          items:
            type: string
        lineItem:
          type: array
          description: Breakdown of the transaction details
          items:
            $ref: '#/components/schemas/LineItem'
        reward:
          $ref: '#/components/schemas/TransactionReward'
          description: Rewards earned with this transaction
        fiAttributes:
          type: array
          description: Array of FI-specific attributes
          items:
            $ref: '#/components/schemas/FiAttribute'
        links:
          $ref: '#/components/schemas/HateoasLinks'
          description: Links (unique to this transaction) used to retrieve images of checks or transaction receipts
      required:
        - accountCategory
      discriminator:
        propertyName: accountCategory
    String255:
      type: string
      title: String 255
      description: String of maximum length 255
      maxLength: 255
...
    DepositTransaction:
      type: object
      title: Deposit Transaction entity
      description: A transaction on a deposit account type
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            transactionType:
              $ref: '#/components/schemas/DepositTransactionType'
              description: 'CHECK, WITHDRAWAL, TRANSFER, POSDEBIT, ATMWITHDRAWAL, BILLPAYMENT, FEE, DEPOSIT, ADJUSTMENT, INTEREST, DIVIDEND, DIRECTDEPOSIT, ATMDEPOSIT, POSCREDIT'
            payee:
              $ref: '#/components/schemas/String255'
              description: Payee name
            checkNumber:
              type: integer
              description: Check number
    DepositTransactionType:
      type: string
      enum:
        - ADJUSTMENT
        - ATMDEPOSIT
        - ATMWITHDRAWAL
        - BILLPAYMENT
        - CHECK
        - DEPOSIT
        - DIRECTDEPOSIT
        - DIVIDEND
        - FEE
        - INTEREST
        - POSCREDIT
        - POSDEBIT
        - TRANSFER
        - WITHDRAWAL
      title: Deposit Transaction Type
      description: The type of a deposit transaction
    InsuranceTransaction:
      type: object
      title: Insurance Transaction entity
      description: An insurance transaction type
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            transactionType:
              $ref: '#/components/schemas/InsuranceTransactionType'
    InsuranceTransactionType:
      type: string
      enum:
        - ADJUSTMENT
        - FEE
        - INTEREST
        - PAYMENT
      title: Insurance Transaction Type
      description: The type of an insurance transaction
    InvestmentTransaction:
      type: object
      title: Investment Transaction entity
      description: Specific transaction information
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            transactionType:
              $ref: '#/components/schemas/InvestmentTransactionType'
              description: 'PURCHASED, SOLD, PURCHASEDTOCOVER, ADJUSTMENT, PURCHASETOOPEN, PURCHASETOCLOSE, SOLDTOOPEN, SOLDTOCLOSE, INTEREST, MARGININTEREST, REINVESTOFINCOME, RETURNOFCAPITAL, TRANSFER, CONTRIBUTION, FEE, OPTIONEXERCISE, OPTIONEXPIRATION, DIVIDEND, DIVIDENDREINVEST, SPLIT, CLOSURE, INCOME, EXPENSE, CLOSUREOPT, INVEXPENSE, JRNLSEC, JRNLFUND, OTHER, DIV, SRVCHG, DEP, DEPOSIT, ATM, POS, XFER, CHECK, PAYMENT, CASH, DIRECTDEP, DIRECTDEBIT, REPEATPMT'
            shares:
              type: number
              description: 'Required for stock, mutual funds. Number of shares (with decimals). Negative numbers indicate securities are being removed from the account'
            faceValue:
              type: number
              description: Cash value for bonds
            price:
              type: number
              description: Unit purchase price
            securityId:
              type: string
              description: Unique identifier of security
            securityIdType:
              $ref: '#/components/schemas/SecurityIdType'
              description: 'CINS, CMC, CME, CUSIP, ISIN, ITSA, NASDAQ, SEDOL, SICC, VALOR, WKN'
            securityType:
              $ref: '#/components/schemas/SecurityType'
              description: 'BOND, DEBT, DIGITALASSET, MUTUALFUND, OPTION, OTHER, STOCK, SWEEP'
            symbol:
              type: string
              description: Ticker symbol
            markup:
              type: number
              description: Portion of unit price that is attributed to the dealer markup
            commission:
              type: number
              description: Transaction commission
            taxes:
              type: number
              description: Taxes on the trade
            fees:
              type: number
              description: Fees applied to the trade
            load:
              type: number
              description: Load on the transaction
            inv401kSource:
              $ref: '#/components/schemas/Inv401kSourceType'
              description: 'Source of money. One of PRETAX, AFTERTAX, MATCH, PROFITSHARING, ROLLOVER, OTHERVEST, OTHERNONVEST'
            confirmationNumber:
              type: string
              description: Confirmation number of the transaction
            fractionalCash:
              type: number
              description: Cash for fractional units (used for stock splits)
            incomeType:
              $ref: '#/components/schemas/IncomeType'
              description: 'Type of investment income. One of CGLONG (capital gains-long term), CGSHORT (capital gains-short term), MISC'
            oldUnits:
              type: number
              description: Number of shares before split
            splitRatioNumerator:
              type: number
              description: Split ratio numerator
            splitRatioDenominator:
              type: number
              description: Split ratio denominator
            newUnits:
              type: number
              description: Number of shares after split
            subAccountSec:
              $ref: '#/components/schemas/SubAccountType'
              description: 'Sub-account security Type. One of CASH, MARGIN, SHORT, OTHERS'
            subAccountFund:
              $ref: '#/components/schemas/SubAccountType'
              description: 'From which account money came in. One of CASH, MARGIN, SHORT, OTHERS'
            loanId:
              type: string
              description: For 401k accounts only. This indicates the transaction was due to a loan or a loan repayment
            loanPrincipal:
              type: number
              description: How much loan pre-payment is principal
            loanInterest:
              type: number
              description: How much loan pre-payment is interest
            payrollDate:
              $ref: '#/components/schemas/DateString'
              description: The date for the 401k transaction was obtained in payroll
            priorYearContrib:
              type: boolean
              description: Indicates this buy was made using prior year's contribution
            withholding:
              type: number
              description: Federal tax withholding
            taxExempt:
              type: boolean
              description: Tax-exempt transaction
            gain:
              type: number
              description: For sales
            stateWithholding:
              type: number
              description: State tax withholding
            penalty:
              type: number
              description: Indicates amount withheld due to a penalty
            runningBalance:
              type: number
              description: Running balance of the position
            unitPrice:
              type: number
              description: 'Price per commonly-quoted unit. Does not include markup/markdown, unitprice. Share price for stocks, mutual funds, and others. Percentage of par for bonds. Per share (not contract) for options'
            units:
              type: number
              description: 'For security-based actions other than stock splits, quantity. Shares for stocks, mutual funds, and others. Face value for bonds. Contracts for options'
            unitType:
              $ref: '#/components/schemas/UnitType'
              description: 'SHARES, CURRENCY'
            transactionReason:
              $ref: '#/components/schemas/TransactionReason'
              description: 'Reason for this transaction; CALL (the debt was called), SELL (the debt was sold), MATURITY (the debt reached maturity)'
            accruedInterest:
              type: number
              description: Accrued interest
            transferAction:
              $ref: '#/components/schemas/TransferDirection'
              description: Transfer direction
            positionType:
              $ref: '#/components/schemas/PositionType'
              description: 'LONG, SHORT'
            digitalUnits:
              type: string
              description: 'Full precision unit number, unlimited digits after decimal point'
            settlementTimestamp:
              $ref: '#/components/schemas/Timestamp'
              description: When the trade settled
...
    IncomeType:
      type: string
      enum:
        - CGLONG
        - CGSHORT
        - MISC
      title: Income Type
      description: The type of income of an investment transaction
...
    LineOfCreditTransaction:
      type: object
      title: Line of Credit Transaction entity
      description: A line of credit transaction
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            transactionType:
              $ref: '#/components/schemas/LineOfCreditTransactionType'
              description: 'CHECK, WITHDRAWAL, PAYMENT, FEE, ADJUSTMENT, INTEREST, PURCHASE'
            checkNumber:
              type: integer
              description: Check number
            paymentDetails:
              $ref: '#/components/schemas/PaymentDetails'
              description: Breakdown of payment details
...
    LoanTransaction:
      type: object
      title: Loan Transaction entity
      description: A transaction on a loan account
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            transactionType:
              $ref: '#/components/schemas/LoanTransactionType'
              description: 'PAYMENT, FEE, ADJUSTMENT, INTEREST'
            paymentDetails:
              $ref: '#/components/schemas/PaymentDetails'
              description: Breakdown of payment details
...
    ActorType:
      type: string
      enum:
        - BATCH
        - USER
      title: FDX API Actor Type
      description: 'Indicates whether a customer is present and has requested the operation (USER), or if a batch job has requested the operation (BATCH)'